Make instantiation of models optional
Upon calling set_spec() in the search space objects, the
graph for that spec is created by default. This change
makes that behaviour optional. Search spaces can now
simply hold the spec for a given architecture, making
it possible to call sample_random_architecture() on
a SearchSpace object multiple times