Make Zen-NAS proxy compatible with search spaces
Zen-NAS now works with NB101, NB201, NB301 and TNB101 search spaces

Note: For autoencoder, normal, and segmentsemantic tasks, being pointwise
tasks, the models have neither global average pooling nor a linear layer.
In this case, there is the question of which feature map to use in place
of the "pre global-average-pooling" feature map. Currently, we use the
input features to the last convolution in the decoder.